Added Thumbnail Selection
Added the opportunity to switch a GalleryView into "Thumbnail Selection
mode". Selection can be indicated by glow and / or an Icon. An example
Icon is included but should be replaced by a good quality Retina icon.

You can see this working in the "Custom Controls Gallery".

How to enable and use it:

[localGallery setIsSelectingView:YES];
[localGallery setSelectingViewHasGlow:YES];
[localGallery setSelectingViewHasIcon:YES];